# swagger-1.0.0-resolved.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Prueba técnica BRM",
    "description": "API para la prueba técnica BRM",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/FLOREZMARLIN1996/brm/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    },
    {
      "url": "https://vl6r2s5z-3000.use2.devtunnels.ms/api",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "description": "Devuelve todos los usuarios registrados",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "Todos los usuarios registrados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/inline_response_200"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Guardar usuario"
        ],
        "summary": "Crear usuario",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Usuario creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_1"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "summary": "Obtener usuario por ID",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Usuario"
        ],
        "summary": "Actualizar usuario por ID",
        "operationId": "updateUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users_id_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_2"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Iniciar sesión de usuario",
        "operationId": "loginUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users_login_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_3"
                }
              }
            }
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": [
          "Productos"
        ],
        "summary": "Obtener todos los productos",
        "operationId": "getAllProducts",
        "responses": {
          "200": {
            "description": "Lista de todos los productos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401_1"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Productos"
        ],
        "summary": "Crear un nuevo producto",
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/products_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Producto creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401_2"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Productos"
        ],
        "summary": "Obtener producto por ID",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Productos"
        ],
        "summary": "Actualizar producto por ID",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/products_id_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Producto actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Productos"
        ],
        "summary": "Eliminar producto por ID",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto eliminado exitosamente"
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/compras": {
      "get": {
        "tags": [
          "Compras"
        ],
        "summary": "Obtener todas las compras",
        "operationId": "getAllCompras",
        "responses": {
          "200": {
            "description": "Lista de todas las compras",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Compras"
        ],
        "summary": "Crear una nueva compra",
        "operationId": "createCompra",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/compras_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Compra creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/compras/{id}": {
      "get": {
        "tags": [
          "Compras"
        ],
        "summary": "Obtener compra por ID",
        "operationId": "getCompraById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compra encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/comprasBatch": {
      "get": {
        "tags": [
          "Compras"
        ],
        "summary": "Obtener compras en lote",
        "operationId": "getComprasBatch",
        "responses": {
          "200": {
            "description": "Lista de compras en lote",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/compras/user/{userId}": {
      "get": {
        "tags": [
          "Compras"
        ],
        "summary": "Obtener compras por ID de usuario",
        "operationId": "getComprasByUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de compras del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/ordenes": {
      "get": {
        "tags": [
          "Órdenes"
        ],
        "summary": "Obtener todas las órdenes",
        "operationId": "getAllOrdenes",
        "responses": {
          "200": {
            "description": "Lista de todas las órdenes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Órdenes"
        ],
        "summary": "Crear una nueva orden",
        "operationId": "createOrden",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ordenes_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Orden creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/ordenesCompra/{compraId}": {
      "get": {
        "tags": [
          "Órdenes"
        ],
        "summary": "Obtener órdenes por ID de compra",
        "operationId": "getOrdenesByCompra",
        "parameters": [
          {
            "name": "compraId",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de órdenes de la compra",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/ordenes/{id}": {
      "delete": {
        "tags": [
          "Órdenes"
        ],
        "summary": "Eliminar orden por ID",
        "operationId": "deleteOrden",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orden eliminada exitosamente"
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "products_body": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "example": "Pepinillo"
          },
          "precio": {
            "type": "number",
            "example": 8
          },
          "disponibles": {
            "type": "number",
            "example": 8
          },
          "lote": {
            "type": "number",
            "example": 2
          }
        }
      },
      "inline_response_401_2": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "noauth"
          }
        }
      },
      "users_login_body": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john_doe@email"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        }
      },
      "ordenes_body": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 100
          },
          "valorunitario": {
            "type": "number",
            "example": 25
          },
          "cantidad": {
            "type": "number",
            "example": 4
          },
          "compra_id": {
            "type": "number",
            "example": 1
          },
          "producto_id": {
            "type": "number",
            "example": 7
          }
        }
      },
      "users_body": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "example": "john_doe"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "email": {
            "type": "string",
            "example": "jon@email.com"
          },
          "phone": {
            "type": "string",
            "example": "3128783274p"
          }
        }
      },
      "users_id_body": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "example": "john_doe"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "email": {
            "type": "string",
            "example": "jon@email.com"
          },
          "phone": {
            "type": "string",
            "example": "3128783274p"
          }
        }
      },
      "compras_body": {
        "type": "object",
        "properties": {
          "descripcion": {
            "type": "string",
            "example": "utiles"
          },
          "total": {
            "type": "number",
            "example": 50
          },
          "user": {
            "type": "number",
            "example": 4
          }
        }
      },
      "inline_response_200_1": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Usuario creado exitosamente"
          }
        }
      },
      "inline_response_200": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "johndoe@example.com"
          }
        }
      },
      "inline_response_200_2": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Usuario actualizado correctamente"
          }
        }
      },
      "inline_response_200_3": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Inicio de sesión exitoso"
          }
        }
      },
      "products_id_body": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "example": "Pepinillo"
          },
          "precio": {
            "type": "number",
            "example": 8
          },
          "disponibles": {
            "type": "number",
            "example": 8
          },
          "lote": {
            "type": "number",
            "example": 2
          }
        }
      },
      "inline_response_200_4": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "ok"
          }
        }
      },
      "inline_response_401": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Token inválido o faltante"
          }
        }
      },
      "inline_response_401_1": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "No auth"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}